requirements.txt
dash==2.18.2
jsonschema==4.24.0
langchain_ollama==0.3.3
matplotlib==3.10.3
numpy==2.3.1
openai==1.93.0
pandas==2.3.0
plotly==6.0.0
protobuf==6.31.1
python-dotenv==1.1.1
scipy==1.16.0
seaborn==0.13.2
streamlit==1.43.1
tqdm==4.67.1
ydata_profiling==4.13.0
